chore: update credential preview screens
