+++ locales/en/index.yml
@@ -498,13 +498,13 @@ profile:
     list:
       wip: In progress
       digitalDomicile: Digital domicile
-      service_contact: 
+      service_contact:
         title: Use of services
         subtitle: Choose who can contact you
       send_email_messages:
         title: Forwarding messages by email
         subtitle: Receive message previews
-      language: 
+      language:
         title: Language
         subtitle: Choose language
       preferred_calendar:
@@ -965,13 +965,13 @@ onboarding:
         notEnrolled:
           text: You need to configure this in your device settings.
           label: How to
-          step1: 
+          step1:
             label: Step 1
             value: Go to your device settings
-          step2: 
+          step2:
             label: Step 2
             value: Set up face or fingerprint
-          step3: 
+          step3:
             label: Step 3
             value: Enable it in IO under Profile > Security
     unavailable:
@@ -979,10 +979,10 @@ onboarding:
       subtitle: IO sends you messages that are relevant to you. Make sure that only you can access the app by setting a code or sign to unlock the device.
       body:
         label: How to
-        step1: 
+        step1:
           label: Step 1
           value: Go to your device settings
-        step2: 
+        step2:
           label: Step 2
           value: Set a code or sign to unlock the device
   shareData:
@@ -3236,12 +3236,13 @@ features:
         tos:
           0: "Per maggiori informazioni, leggi"
           1: l’informativa Privacy e i Termini e Condizioni d’uso
+      eidPreview:
+        title: "Stiamo aggiungendo i seguenti dati al Portafoglio:"
+        actions:
+          primary: Continua
+          secondary: Annulla
       credentialPreview:
-        title: "{{credential}}: ecco l'anteprima"
-        banner:
-          title: "Qualcosa non torna?"
-          content: "Se ci sono errori nei dati o vuoi capire meglio cosa significano, contatta l’assistenza."
-          actionTitle: "Contatta l'assistenza"
+        title: "{{credential}}: ecco l'anteprima dei tuoi dati"
         bottomSheet:
           about:
             title: "Chi è?"
@@ -3249,9 +3250,6 @@ features:
           data:
             title: "Protezione dei dati"
             subtitle: "I tuoi dati non vengono salvati su questo dispositivo, ma conservati nella Wallet Instance e potrai decidere di revocarli in qualsiasi momento."
-        toast:
-          success: "Tessera aggiunta!"
-          failure: "Non siamo riusciti ad aggiungere questa tessera, riprova più tardi."
         actions:
           primary: Aggiungi al portafoglio
           secondary: Annulla
